refactor(keypair): move logic to a module
move bulk of testing there, Verify still should have a keypair
